# backend/internal/handlers/rest/v1/model_game.go
/*
 * Ibis
 *
 * API for Ibis
 *
 * API version: 1.0.0
 * Contact: holmes89@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package v1

type Game struct {
	Id string `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Path string `json:"path,omitempty"`

	Image string `json:"image,omitempty"`

	Platform string `json:"platform,omitempty"`

	Description string `json:"description,omitempty"`
}

// AssertGameRequired checks if the required fields are not zero-ed
func AssertGameRequired(obj Game) error {
	return nil
}

// AssertRecurseGameRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Game (e.g. [][]Game), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseGameRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aGame, ok := obj.(Game)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertGameRequired(aGame)
	})
}